+++ client/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
+### Added
+
+- diff in transport view
+
+### Changed
+
+- some icons, borrowed from gitlens
+- decentralised command definitions
+
 ## 0.5.27 2019-03-30
 
 ### Added